requirements.txt
google-api-python-client==2.*
google-auth-httplib2==0.*
google-auth-oauthlib==1.*